+++ src/mass_driver/models/repository.py
@@ -13,7 +13,7 @@ RepoUrl = str
 """A repo's clone URL, either git@github.com format or local filesystem path"""
 
 
-class Repo(BaseModel):
+class SourcedRepo(BaseModel):
     """A repository as-discovered, pre-cloning, along with any Sourced metadata"""
 
     clone_url: RepoUrl
@@ -28,7 +28,7 @@ class Repo(BaseModel):
     """Arbitrary data dict from Source"""
 
 
-class ClonedRepo(Repo):
+class ClonedRepo(SourcedRepo):
     """A repository after it has been successfully cloned, branch configured"""
 
     cloned_path: DirectoryPath
@@ -37,15 +37,15 @@ class ClonedRepo(Repo):
     """The name of the currently checked out branch"""
 
 
-IndexedRepos = dict[RepoID, Repo]
-"""A "list" of repositories, but indexed by repo ID (Repo.repo_id)"""
+IndexedRepos = dict[RepoID, SourcedRepo]
+"""A "list" of repositories, but indexed by repo ID (SourcedRepo.repo_id)"""
 
 IndexedClonedRepos = dict[RepoID, ClonedRepo]
 """A "list" of cloned repositories, but indexed by repo ID (ClonedRepo.repo_id)"""
 
 
 class Source(BaseSettings):
-    """Base class for Sources of Repo, on which to apply patching or scan"""
+    """Base class for Sources of SourcedRepo, on which to apply patching or scan"""
 
     def discover(self) -> IndexedRepos:
         """Discover a list of repositories"""